requirements.txt
pyyaml>=6.0
pytest>=7.0
matplotlib>=3.8.0
